end project

+++ AlkinanaPharmaManagment.Application/Products/Delete/DeleteProductCommandHandler.cs
@@ -1,11 +1,7 @@
 ﻿using AlkinanaPharmaManagment.Application.Abstractions.Messaging;
-using AlkinanaPharmaManagment.Domain.Repositories;
+using AlkinanaPharmaManagment.Application.Exceptions;
+using AlkinanaPharmaManagment.Domain.Entities.Products.Events;
 using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AlkinanaPharmaManagment.Application.Products.Delete
 {
@@ -13,6 +9,15 @@ namespace AlkinanaPharmaManagment.Application.Products.Delete
     {
         public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
+            var product = await productRepository.GetFullProductById(request.productId);
+
+            if (product is null)
+            {
+                throw new NotFoundException(nameof(product),request.productId);
+            }
+
+            product.Raise(new ProductDeletedEvent(product));
+
             await productRepository.DeleteAsync(request.productId);
 
             await productRepository.SaveChangeAsync();
